@@ -3,7 +3,7 @@
   "description": "PayGate PayWeb Payment Gateway for Magento 2",
   "type": "magento2-module",
   "minimum-stability": "stable",
-  "license": "proprietary",
+  "license": "GPL-3.0",
   "authors": [
     {
       "name": "App Inlet (Pty) Ltd",
@@ -11,10 +11,14 @@
     }
   ],
   "require": {
-    "php": "~7.4.0"
+    "php": "^8.1",
+    "ext-curl": "*",
+    "ext-http": "*"
   },
   "autoload": {
-    "files": ["paygate/paygate-payweb-gateway/registration.php"],
+    "files": [
+      "paygate/paygate-payweb-gateway/registration.php"
+    ],
     "psr-4": {
       "PayGate\\PayWeb\\": "paygate/paygate-payweb-gateway"
     }
