@@ -1,4 +1,10 @@
 <?php
+/** @noinspection PhpUnused */
+
+/** @noinspection PhpUndefinedNamespaceInspection */
+
+/** @noinspection PhpUndefinedNamespaceInspection */
+
 /*
  * Copyright (c) 2022 PayGate (Pty) Ltd
  *
@@ -21,8 +27,9 @@ class CardRenderer extends AbstractCardRenderer
      *
      * @return boolean
      * @since 100.1.0
+     * @noinspection PhpUnused
      */
-    public function canRender(PaymentTokenInterface $token)
+    public function canRender(PaymentTokenInterface $token): bool
     {
         return $token->getPaymentMethodCode() === "paygate";
     }
@@ -30,8 +37,10 @@ class CardRenderer extends AbstractCardRenderer
     /**
      * @return string
      * @since 100.1.0
+     * @noinspection PhpUnused
+     * @noinspection PhpUndefinedMethodInspection
      */
-    public function getNumberLast4Digits()
+    public function getNumberLast4Digits(): string
     {
         return $this->getTokenDetails()['maskedCC'];
     }
@@ -39,8 +48,10 @@ class CardRenderer extends AbstractCardRenderer
     /**
      * @return string
      * @since 100.1.0
+     * @noinspection PhpUnused
+     * @noinspection PhpUndefinedMethodInspection
      */
-    public function getExpDate()
+    public function getExpDate(): string
     {
         return $this->getTokenDetails()['expirationDate'];
     }
@@ -48,8 +59,10 @@ class CardRenderer extends AbstractCardRenderer
     /**
      * @return string
      * @since 100.1.0
+     * @noinspection PhpUnused
+     * @noinspection PhpUndefinedMethodInspection
      */
-    public function getIconUrl()
+    public function getIconUrl(): string
     {
         return $this->getIconForType($this->getTokenDetails()['type'])['url'];
     }
@@ -57,8 +70,10 @@ class CardRenderer extends AbstractCardRenderer
     /**
      * @return int
      * @since 100.1.0
+     * @noinspection PhpUnused
+     * @noinspection PhpUndefinedMethodInspection
      */
-    public function getIconHeight()
+    public function getIconHeight(): int
     {
         return $this->getIconForType($this->getTokenDetails()['type'])['height'];
     }
@@ -66,8 +81,10 @@ class CardRenderer extends AbstractCardRenderer
     /**
      * @return int
      * @since 100.1.0
+     * @noinspection PhpUnused
+     * @noinspection PhpUndefinedMethodInspection
      */
-    public function getIconWidth()
+    public function getIconWidth(): int
     {
         return $this->getIconForType($this->getTokenDetails()['type'])['width'];
     }
